% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{pp_assign_latex}
\alias{pp_assign_latex}
\title{Layout padrão da assinatura do relatório emitido em 'rmarkdown' do tipo 'latex'}
\usage{
pp_assign_latex(x, y, z)
}
\arguments{
\item{x}{Primeiro item da assinatura, podendo ser o nome}

\item{y}{Segundo item da assinatura, podendo ser o cargo ou setor}

\item{z}{Terceiro item da assinatura, podendo ser o local e data}
}
\value{
Layout padrão da assinatura
}
\description{
Layout padrão da assinatura do relatório emitido em 'rmarkdown' do tipo 'latex'
}
