% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pp_complete}
\alias{pp_complete}
\title{Função completa que aplica as funções de coleta e tratamento dos dados}
\usage{
pp_complete(
  df,
  vars_group,
  available = FALSE,
  unselect = NULL,
  total_vars_group,
  total_name = NULL,
  total_all = TRUE,
  position = "down",
  vars_suppress = NULL,
  per_num,
  per_den,
  per_nsmall = 1,
  per_format = F,
  per_diff = F,
  arrange
)
}
\arguments{
\item{df}{'data.frame' da base de dados}

\item{vars_group}{Variáveis do 'data.frame' a serem agrupadas por meio da
função 'pp_execucao'}

\item{available}{Considera detalhamento do 'CREDITO_DISPONIVEL' se 'TRUE' a
ser considerada na função 'pp_execucao'}

\item{unselect}{Variáveis do 'data.frame' a serem desconsideradas na
função 'pp_execucao'}

\item{total_vars_group}{Variáveis do 'data.frame' a serem agrupadas por meio
da função 'pp_subtotal'}

\item{total_name}{Nome da linha que conterá o somatório}

\item{total_all}{Considera soma total por colunas numéricas se 'TRUE'}

\item{position}{Posição das linhas subtotais no 'data.frame', podendo ser
'up' para acima das variáveis agrupadas ou 'down' para baixo dessas variáveis}

\item{vars_suppress}{Variáveis a serem suprimidas}

\item{per_num}{Numerador da fração, podendo ser uma ou mais colunas de classe 
numérica de determinado 'data.frame' ou vetor de classe numérica, caso o 
parâmetro 'df' não seja informado.}

\item{per_den}{Denominador da fração, podendo uma coluna de classe 
numérica de determinado 'data.frame' ou vetor de classe numérica, caso
o parâmetro 'df' não seja informado.}

\item{per_nsmall}{Número de casas decimais após a vírgula, correspondente ao
parâmetro 'nsmall' da função 'base::format'}

\item{per_format}{Opção verdadeira 'TRUE' ou falsa 'FALSE', caso necessária
a formatação do resultado final da fração}

\item{per_diff}{Calcula a partir da diferenca}

\item{arrange}{Ordenar por variáveis indicadas em 'arrange'}
}
\value{
'data.frame' com o mês a mês da execução orçamentária e financeira
}
\description{
Função completa que aplica as funções de coleta e tratamento dos dados
}
