% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{pp_content_latex}
\alias{pp_content_latex}
\title{Layout padrão da tabela de conteúdo para relatórios em 'rmarkdown' 
do tipo 'latex'}
\usage{
pp_content_latex(
  bg_color = "ffffff",
  border_color = "ffffff",
  border_width = 0.1
)
}
\arguments{
\item{bg_color}{Cor do fundo da tabela de conteúdo}

\item{border_color}{Cor da borda da tabela de conteúdo}

\item{border_width}{Tamanho da linha da borda da tabela de conteúdo}
}
\value{
Layout padrão da tabela de conteúdo
}
\description{
Layout padrão da tabela de conteúdo para relatórios em 'rmarkdown' 
do tipo 'latex'
}
