% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pp_database2}
\alias{pp_database2}
\title{Tratamento da base de dados (Novo)}
\usage{
pp_database2(path, database)
}
\arguments{
\item{path}{Caminho do arquivo base no formato 'xlsx'}

\item{database}{Base de dados correspondente ao arquivo 'xlsx', podendo ser:
- tg_nota_credito
- tg_nota_dotacao
- tg_nota_empenho
- tg_orcamento
- tg_receita_propria
- sp_pessoal}
}
\value{
Dado tratado
}
\description{
Tratamento da base de dados (Novo)
}
