% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pp_execucao}
\alias{pp_execucao}
\title{Gera dados de execução orçamentária e financeira}
\usage{
pp_execucao(df, vars_group, available = F, unselect = NULL)
}
\arguments{
\item{df}{'data.frame' da base de dados 'tg_orcamento'}

\item{vars_group}{Variáveis do 'data.frame' a serem agrupadas}

\item{available}{Considera detalhamento do 'CREDITO_DISPONIVEL' se 'TRUE'}

\item{unselect}{Variáveis do 'data.frame' a serem desconsideradas}
}
\value{
'data.frame' com a execução orçamentária e financeira
}
\description{
Gera dados de execução orçamentária e financeira
}
