% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pp_execucao_mes}
\alias{pp_execucao_mes}
\title{Gera dados mensais de execução orçamentária e financeira}
\usage{
pp_execucao_mes(
  df,
  vars_group,
  available = F,
  unselect = NULL,
  cumsum_num = F,
  percent = F,
  ...
)
}
\arguments{
\item{df}{'data.frame' da base de dados 'tg_orcamento'}

\item{vars_group}{Variáveis do 'data.frame' a serem agrupadas}

\item{available}{Considera detalhamento do 'CREDITO_DISPONIVEL' se 'TRUE'}

\item{unselect}{Variáveis do 'data.frame' a serem desconsideradas}

\item{cumsum_num}{Considera soma acumulada por colunas numéricas se 'TRUE'}

\item{percent}{Acrescenta variáveis a partir da função 'pp_percent' se 'TRUE.
Requer detalhamento em '...'}

\item{...}{Argumentos da função 'pp_percent'}
}
\value{
'data.frame' com o mês a mês da execução orçamentária e financeira
}
\description{
Gera dados mensais de execução orçamentária e financeira
}
