% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{pp_flextable}
\alias{pp_flextable}
\title{Tabela padrão do tipo 'flextable' a ser aplicado nos relatórios em 'rmarkdown'}
\usage{
pp_flextable(
  x,
  caption,
  title,
  title_group,
  merge,
  group,
  height,
  width,
  font_size = 10,
  ...
)
}
\arguments{
\item{x}{Data.frame para conversão em tabela do tipo 'flextable'}

\item{caption}{Título da tabela}

\item{title}{Novos títulos para as colunas da tabela. Se ausente 'NULL', retorna
nomes das colunas do 'data.frame' corrigido pela função 'pp_format_str'}

\item{title_group}{Cria um título mesclado para agrupamento de colunas especificados
no argumento}

\item{merge}{Colunas e/ou linhas a serem mescladas}

\item{group}{Agrupamento pela função 'as_grouped_data'}

\item{height}{Altura da linha em pixels}

\item{width}{Tamanho da coluna em pixels}

\item{font_size}{Tamanho da fonte}

\item{...}{Outros argumentos para a função 'pp_theme'}
}
\value{
Tabela flextable 'html' ou 'latex'
}
\description{
Tabela padrão do tipo 'flextable' a ser aplicado nos relatórios em 'rmarkdown'
}
