% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{pp_flextheme}
\alias{pp_flextheme}
\title{Tema padrão para a tabela do tipo 'flextable' a ser aplicado nos relatórios
'rmarkdown'}
\usage{
pp_flextheme(
  x,
  bg_title = "#336699",
  border_color = "#d9d9d9",
  bg_row,
  align_col
)
}
\arguments{
\item{x}{Objeto 'flextable'}

\item{bg_title}{Cor do fundo para o título}

\item{border_color}{Cor da borda}

\item{bg_row}{Cor do fundo das linhas da tabela}

\item{align_col}{Permite definir o alinhamento para cada coluna}
}
\value{
Tema padrão para a tabela
}
\description{
Tema padrão para a tabela do tipo 'flextable' a ser aplicado nos relatórios
'rmarkdown'
}
