% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_format_month}
\alias{pp_format_month}
\title{Convers<c3><a3>o de m<c3><aa>s num<c3><a9>rico do TG para sigla do m<c3><aa>s}
\usage{
pp_format_month(x)
}
\arguments{
\item{x}{M<c3><aa>s ou meses em formato num<c3><a9>rico (1, 2, 3, 01, 02, 03).}
}
\value{
M<c3><aa>s ou meses no formato texto (Jan, Fev, Mar, Jan, Fev, Mar)
}
\description{
Convers<c3><a3>o de m<c3><aa>s num<c3><a9>rico do TG para sigla do m<c3><aa>s
}
