% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{pp_info_latex}
\alias{pp_info_latex}
\title{Layout padrão da tabela de informações básicas sobre o relatório em
'rmarkdown' do tipo 'latex'}
\usage{
pp_info_latex(x, border_color = "ffffff", border_width = 0.1)
}
\arguments{
\item{x}{Vetor de itens para a tabela}

\item{border_color}{Cor da borda da tabela de conteúdo}

\item{border_width}{Tamanho da linha da borda da tabela de conteúdo}
}
\value{
Layout padrão da tabela de informações básicas
}
\description{
Layout padrão da tabela de informações básicas sobre o relatório em
'rmarkdown' do tipo 'latex'
}
