% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_last_day}
\alias{pp_last_day}
\title{<c3><9a>ltimo dia de determinada refer<c3><aa>ncia mensal ou anual}
\usage{
pp_last_day(x, by = "month", week_start = 1)
}
\arguments{
\item{x}{Data de refer<c3><aa>ncia no formato 'yyyy-mm-dd'}

\item{by}{Unidade da fun<c3><a7><c3><a3>o 'round_date' do pacote lubridate, 
podendo ser 'day', 'week', 'month' ou 'year'}

\item{week_start}{In<c3><ad>cio da semana. Vari<c3><a1>vel a ser utilizada
caso o par<c3><a2>metro 'by' seja igual a 'month'}
}
\value{
<c3><9a>ltimo dia de determinada refer<c3><aa>ncia mensal ou anual
}
\description{
<c3><9a>ltimo dia de determinada refer<c3><aa>ncia mensal ou anual
}
