% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pp_orcamento}
\alias{pp_orcamento}
\title{Extração de dados a partir de referenciais padronizados do orçamento}
\usage{
pp_orcamento(
  df,
  area = NULL,
  co_uge = NULL,
  co_ugr = NULL,
  co_ug_by = "ou",
  rap = T
)
}
\arguments{
\item{df}{'data.frame' da base de dados 'tg_orcamento'}

\item{area}{Área gestora do orçamento, conforme coluna AREA criada pela
função 'pp_database' para a base de dados 'tg_orcamento'}

\item{co_uge}{Código da Unidade Gestora Executora (UGE)}

\item{co_ugr}{Código da Unidade Gestora Responsável (UGR)}

\item{co_ug_by}{Filtrar UGE e/ou UGR}

\item{rap}{Desconsidera restos a pagar (RAP) se 'FALSE'}
}
\value{
'data.frame' do orçamento
}
\description{
Extração de dados a partir de referenciais padronizados do orçamento
}
