% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_progress2}
\alias{pp_progress2}
\title{Barra de progresso impressa no console}
\usage{
pp_progress2(etapa, total, label = "Progresso")
}
\arguments{
\item{etapa}{Valor num<c3><a9>rico da etapa}

\item{total}{Valor num<c3><a9>rico do total}

\item{label}{Texto a ser exibido na mensagem de progresso}
}
\value{
Barra de progresso.
}
\description{
Barra de progresso impressa no console
}
