% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{pp_title_latex}
\alias{pp_title_latex}
\title{Layout padrão para título de relatórios em 'rmarkdown' do tipo 'latex'}
\usage{
pp_title_latex(
  title = "",
  subtitle = "",
  bg_color = "#2f9e41",
  font_color = "#ffffff",
  logo
)
}
\arguments{
\item{title}{Título do relatório}

\item{subtitle}{Subtítulo do relatório}

\item{bg_color}{Cor do fundo do título}

\item{font_color}{Cor da fonte a ser aplicada ao título 'title'  e subtítulo 'subtitle'}

\item{logo}{Caminho para o arquivo de logotipo}
}
\value{
Layout padrão do título
}
\description{
Layout padrão para título de relatórios em 'rmarkdown' do tipo 'latex'
}
