% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_total_fixed}
\alias{pp_total_fixed}
\title{Soma total de determinada coluna de um 'data.frame'}
\usage{
pp_total_fixed(df, fixed)
}
\arguments{
\item{df}{Objeto do tipo 'data.frame'.}

\item{fixed}{Colunas do 'data.frame' fixas, a serem desconsideradas
em determinado cálculo somatório do total.}
}
\value{
Soma total por colunas de um 'data.frame'
}
\description{
Soma total de determinada coluna de um 'data.frame'
}
