% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_format_month}
\alias{pp_format_month}
\title{Conversão de mês numérico do TG para sigla do mês}
\usage{
pp_format_month(x)
}
\arguments{
\item{x}{Mês ou meses em formato numérico (1, 2, 3, 01, 02, 03).}
}
\value{
Mês ou meses no formato texto (Jan, Fev, Mar, Jan, Fev, Mar)
}
\description{
Conversão de mês numérico do TG para sigla do mês
}
