% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_format_str}
\alias{pp_format_str}
\title{Formatação padrão de uma 'string' para tratamento de dados}
\usage{
pp_format_str(x, enc = "Latin-ASCII", sep = "_", str_to = "upper")
}
\arguments{
\item{x}{Fragmento de texto.}

\item{enc}{Codificação de texto padrão.}

\item{sep}{Separador a ser utilizado em substituição ao espaço ' '.}

\item{str_to}{Formatação final compatível com as funções
do pacote 'stringr': 
- upper (str_to_upper) para tornar todas as letras maiúsculas
- lower (str_to_lower) para tornar todas as letras minúsculas
- sentence (str_to_sentence) para a primeira letra maiúscula e as 
demais minúsculas
- title (str_to_title) tornar maiúscia a primeira letra de cada 
palavra que compôem o fragmento de texto}
}
\value{
Fragmento de texto tratado.
}
\description{
Formatação padrão de uma 'string' para tratamento de dados
}
