% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pp_group_sum}
\alias{pp_group_sum}
\title{Soma e agrupamento de vari치veis}
\usage{
pp_group_sum(df, vars_group, vars_sum)
}
\arguments{
\item{df}{'data.frame'}

\item{vars_group}{Vari치veis do 'data.frame' a serem agrupadas}

\item{vars_sum}{Vari치veis do 'data.frame' a serem somadas}
}
\value{
'data.frame' agrupado
}
\description{
Soma e agrupamento de vari치veis
}
