% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_interval}
\alias{pp_interval}
\title{Intervalo entre datas}
\usage{
pp_interval(start, end = Sys.Date(), by = "day", position = "first")
}
\arguments{
\item{start}{Data de início do intervalo, no formato 'yyyy-mm-dd'}

\item{end}{Data fim do intervalo, no formato 'yyyy-mm-dd'}

\item{by}{Unidade da função 'round_date' do pacote lubridate, 
podendo ser 'day', 'week', 'month' ou 'year'}

\item{position}{Caso preenchido, retorna o primeiro 'first' ou
o último 'last' dia do intervalo para os parâmetros de 'by' 
'month' ou 'year'}
}
\value{
Intervalo entre datas
}
\description{
Intervalo entre datas
}
