% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_last_day}
\alias{pp_last_day}
\title{Último dia de determinada referência mensal ou anual}
\usage{
pp_last_day(x, by = "month", week_start = 1)
}
\arguments{
\item{x}{Data de referência no formato 'yyyy-mm-dd'}

\item{by}{Unidade da função 'round_date' do pacote lubridate, 
podendo ser 'day', 'week', 'month' ou 'year'}

\item{week_start}{Início da semana. Variável a ser utilizada
caso o parâmetro 'by' seja igual a 'month'}
}
\value{
Último dia de determinada referência mensal ou anual
}
\description{
Último dia de determinada referência mensal ou anual
}
