% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_list_rda}
\alias{pp_list_rda}
\title{Importa e lista objetos de determinado arquivo '.rda'}
\usage{
pp_list_rda(x)
}
\arguments{
\item{x}{Arquivo de extensão '.rda' para importação dos dados.
Aceita caminho completo.}
}
\value{
Arquivos '.rda' em formato de lista
}
\description{
Importa e lista objetos de determinado arquivo '.rda'
}
