% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_percent}
\alias{pp_percent}
\title{Calculo percentual}
\usage{
pp_percent(df, num = NULL, den = NULL, nsmall = 1, format = FALSE)
}
\arguments{
\item{df}{Objeto do tipo 'data.frame', o que requer colunas para os 
parâmetros 'num' e 'den'. Pode ser dispensado caso os parâmetros 
'num' e 'den' sejam vetores.}

\item{num}{Numerador da fração, podendo ser uma ou mais colunas de classe 
numérica de determinado 'data.frame' ou vetor de classe numérica, caso o 
parâmetro 'df' não seja informado.}

\item{den}{Denominador da fração, podendo uma coluna de classe 
numérica de determinado 'data.frame' ou vetor de classe numérica, caso
o parâmetro 'df' não seja informado.}

\item{nsmall}{Número de casas decimais após a vírgula, correspondente ao
parâmetro 'nsmall' da função 'base::format'}

\item{format}{Opção verdadeira 'TRUE' ou falsa 'FALSE', caso necessária
a formatação do resultado final da fração}
}
\value{
Vetor 'vector' ou coluna 'column' do tipo numérica ou textual
}
\description{
Calculo percentual
}
