% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pp_progress}
\alias{pp_progress}
\title{Barra de progresso impressa no console}
\usage{
pp_progress(x, msg = "", bar = "-", time = F)
}
\arguments{
\item{x}{Valor numérico da posição na barra de progresso, de 1 a 100.}

\item{msg}{Mensagem de saída}

\item{bar}{Caracter símbolo da barra de progresso}

\item{time}{Inclui a hora atual se 'TRUE'}
}
\value{
Barra de progresso.
}
\description{
Barra de progresso impressa no console
}
