% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pp_subtotal}
\alias{pp_subtotal}
\title{Acrescenta linhas de subtotais e total em 'data.frame'}
\usage{
pp_subtotal(
  df,
  vars_group,
  total_name,
  total_all = T,
  position = "down",
  vars_suppress = NULL
)
}
\arguments{
\item{df}{'data.frame' da base de dados}

\item{vars_group}{Variáveis do 'data.frame' a serem agrupadas}

\item{total_name}{Nome da linha que conterá o somatório}

\item{total_all}{Considera soma total por colunas numéricas se 'TRUE'}

\item{position}{Posição das linhas subtotais no 'data.frame', podendo ser
'up' para acima das variáveis agrupadas ou 'down' para baixo dessas variáveis}

\item{vars_suppress}{Variáveis a serem suprimidas}
}
\value{
'data.frame' com o mês a mês da execução orçamentária e financeira
}
\description{
Acrescenta linhas de subtotais e total em 'data.frame'
}
